requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
redis==5.0.3
pika==1.3.2
python-dotenv==1.0.1